requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-openai
streamlit
langchain_groq
langserve
validators==0.28.1
youtube_transcript_api
unstructured
pytube
crewai
crewai_tools
pandas
numpy